+++ src/Symfony/Component/Config/Definition/Builder/TreeBuilder.php
@@ -29,7 +29,8 @@ class TreeBuilder implements NodeParentInterface
         if (null === $name) {
             @trigger_error('A tree builder without a root node is deprecated since Symfony 4.2 and will not be supported anymore in 5.0.', E_USER_DEPRECATED);
         } else {
-            $this->root($name, $type, $builder);
+            $builder = $builder ?: new NodeBuilder();
+            $this->root = $builder->node($name, $type)->setParent($this);
         }
     }
 
@@ -43,9 +44,13 @@ class TreeBuilder implements NodeParentInterface
      * @return ArrayNodeDefinition|NodeDefinition The root node (as an ArrayNodeDefinition when the type is 'array')
      *
      * @throws \RuntimeException When the node type is not supported
+     *
+     * @deprecated since Symfony 4.3, pass the root name to the constructor instead
      */
     public function root($name, $type = 'array', NodeBuilder $builder = null)
     {
+        @trigger_error(sprintf('The "%s()" method called for the "%s" configuration is deprecated since Symfony 4.3, pass the root name to the constructor instead.', __METHOD__, $name), E_USER_DEPRECATED);
+
         $builder = $builder ?: new NodeBuilder();
 
         return $this->root = $builder->node($name, $type)->setParent($this);