+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * deprecated using environment variables with `cannotBeEmpty()` if the value is validated with `validate()`
  * made `Resource\*` classes final and not implement `Serializable` anymore
+ * deprecated the `root()` method in `TreeBuilder`, pass the root node information to the constructor instead
 
 4.2.0
 -----